+++ src/main/java/racingcar/controller/GameController.java
@@ -1,0 +1,4 @@
+package racingcar.controller;
+
+public class GameController {
+}
