+++ src/main/java/racingcar/domain/RacingGame.java
@@ -1,6 +1,7 @@
 package racingcar;
 
 import camp.nextstep.edu.missionutils.Randoms;
+import racingcar.domain.Car;
 
 import java.util.ArrayList;
 import java.util.List;
